+++ mitra_models/src/users/queries.rs
@@ -18,7 +18,12 @@ use crate::database::{
 };
 use crate::profiles::{
     queries::create_profile,
-    types::{DbActorProfile, MentionPolicy, ProfileCreateData},
+    types::{
+        DbActorProfile,
+        MentionPolicy,
+        ProfileCreateData,
+        WebfingerHostname,
+    },
 };
 
 use super::types::{
@@ -141,7 +146,7 @@ pub async fn create_user(
     // Create profile
     let profile_data = ProfileCreateData {
         username: user_data.username.clone(),
-        hostname: None,
+        hostname: WebfingerHostname::Local,
         display_name: None,
         bio: None,
         avatar: None,
@@ -770,7 +775,7 @@ mod tests {
         let db_client = &mut create_test_database().await;
         let profile_data = ProfileCreateData {
             username: "test".to_string(),
-            hostname: None,
+            hostname: WebfingerHostname::Unknown,
             public_keys: vec![DbActorKey::default()],
             actor_json: Some(DbActor {
                 id: "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor".to_string(),