+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Change poll max expiration time to avoid integer overflow in Mastodon clients.
 - Preserve order of mentions in local posts.
+- Fixed violation of data integrity when portable account is updated.
 
 ## [3.15.0] - 2025-01-24